Updated akka-examples
 - updated sbt command
 - changed package name
 - Corrected the test code according to the latest notation of specs2
