+++ 2.6/async/akka-examples/src/main/scala/org/scalatra/example/FutureController.scala
@@ -1,4 +1,4 @@
-package com.example.app
+package org.scalatra.example
 
 import akka.actor.ActorSystem
 import dispatch._
