+++ 2.6/async/akka-examples/src/test/scala/org/scalatra/example/FutureControllerSpec.scala
@@ -1,0 +1,19 @@
+package org.scalatra.example
+
+import org.scalatra.test.specs2._
+import akka.actor.{ActorSystem}
+
+class FutureControllerSpec extends ScalatraSpec { def is = s2"""
+  GET / on FutureController
+    should return status 200 $root
+"""
+
+  val system = ActorSystem()
+
+  addServlet(new FutureController(system), "/*")
+
+  def root = get("/") {
+    status must_== 200
+  }
+
+}
